+++ docs/changelog.md
@@ -1,6 +1,11 @@
 # Changelog:
 [Semantic Versioning](https://semver.org/)
 
+### v1.8.0
+- Improved documentation
+- Renamed "formatted" to "converted"
+- Added ::fromArray(array $items) static fabric method for **_Wmsamolet\PhpObjectCollections\AbstractCollection_**
+
 ### v1.7.1
 - Improved documentation
 - Fix remove offsetKey(...) return type
@@ -26,7 +31,7 @@
 - Added ->getConvertKeyCallback() method for **_Wmsamolet\PhpObjectCollections\TypedCollection_**
 - Added ->setConvertKeyCallback(...) method for **_Wmsamolet\PhpObjectCollections\TypedCollection_**
 - Added $convertValueCallback __constructor() argument for **_Wmsamolet\PhpObjectCollections\TypedCollection_**
-- Added ::fromIterator(Traversable $iterator) fabric method for **_Wmsamolet\PhpObjectCollections\AbstractCollection_**
+- Added ::fromIterator(Traversable $iterator) static fabric method for **_Wmsamolet\PhpObjectCollections\AbstractCollection_**
  
 ### v1.4.0
 - Added collectionKeyType() method for **_Wmsamolet\PhpObjectCollections\AbstractTypedCollection_**